requirements.txt
-i https://pypi.org/simple
--extra-index-url https://test.pypi.org/simple/
-e .
appdirs==1.4.3
atomicwrites==1.3.0
attrs==19.1.0
autopep8==1.4.4
bcbio-gff==0.6.6
beautifulsoup4==4.7.1
biopython==1.73
bioservices==1.6.0
certifi==2019.3.9
chardet==3.0.4
click==7.0
colorama==0.4.1
colorlog==4.0.2
coverage==4.5.3
coveralls==1.8.0
docopt==0.6.2
easydev==0.9.37
entrypoints==0.3
flake8==3.7.7
gevent==1.4.0
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
grequests==0.3.0
idna==2.8
importlib-metadata==0.17
lxml==4.3.4
mccabe==0.6.1
more-itertools==7.0.0 ; python_version > '2.7'
numpy==1.16.4
packaging==19.0
pandas==0.24.2
pexpect==4.7.0
pluggy==0.12.0
ptyprocess==0.6.0
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.0
pytest-cov==2.7.1
pytest==4.6.2
python-dateutil==2.8.0
pytz==2019.1
requests-cache==0.5.0
requests==2.22.0
six==1.12.0
soupsieve==1.9.1
suds-jurko==0.6
urllib3==1.25.3
wcwidth==0.1.7
wrapt==1.11.1
xmltodict==0.12.0
zipp==0.5.1
tqdm
